backend/requirements.txt
# FastAPI and server
fastapi>=0.128.0
uvicorn>=0.23.0
python-multipart>=0.0.21

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0

# Image Processing
Pillow>=10.0.0

# Data validation
pydantic>=2.0.0

# Testing
pytest>=9.0.0
httpx>=0.28.0

# Utilities
python-json-logger>=2.0.0
